Clarify the `generators` API.
The `GeneratorsView` struct was originally intended to be a view into the
"full" (m,n) generators for a specific proof share.  But the
`Generators::all()` method gave a `GeneratorsView` that covered all of the
generators. This meant that `GeneratorsView` ended up having a duplicate role
to `&Generators`, and that it wasn't possible to tell whether a
`GeneratorsView` was for a specific share or was being used like a
`&Generators`.

This commit drops the `Generators::all()` method in order to fix this
inconsistency.